backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
scikit-learn
requests
cryptography
SpeechRecognition
openai-whisper
python-dotenv
